// nodes/Commercetools/generated/properties.ts
export const generatedProperties: INodeProperties[] = [
// ...
  {
    "displayName": "Operation",
    "name": "operation",
    "type": "options",
    "noDataExpression": true,
    "displayOptions": {
      "show": {
        "resource": [
          "products"
        ]
      }
    },
// ...
  },
// ...
  {
    "displayName": "Product ID",
    "name": "resourceId",
    "type": "string",
    "default": "",
    "required": true,
    "displayOptions": {
      "show": {
        "resource": [
          "products"
        ],
        "operation": [
          "uploadProductImage",
          "queryProductSelectionsForProductByProductId",
          "getProductById",
          "checkIfProductExistsById",
          "updateProductById",
          "deleteProductById"
        ]
      }
    }
  },
// ...
];